# backend/tests/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentHub Multi-Agent Chatbot Framework",
    "description": "Production-ready multi-tenant chatbot framework using LangChain 0.3+",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/tenants/{tenant_id}/knowledge": {
      "post": {
        "tags": [
          "admin-knowledge",
          "admin-knowledge"
        ],
        "summary": "Ingest Documents",
        "description": "Ingest documents into tenant's knowledge base.\n\nThis creates a tenant-specific ChromaDB collection and adds documents\nfor later retrieval by AgentAnalysis via RAG.\n\nRequires admin role in JWT.",
        "operationId": "ingest_documents_api_admin_tenants__tenant_id__knowledge_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-knowledge",
          "admin-knowledge"
        ],
        "summary": "Delete Documents",
        "description": "Delete documents from tenant's knowledge base.\n\nRequires admin role in JWT.",
        "operationId": "delete_documents_api_admin_tenants__tenant_id__knowledge_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Document Ids"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tenants/{tenant_id}/knowledge/stats": {
      "get": {
        "tags": [
          "admin-knowledge",
          "admin-knowledge"
        ],
        "summary": "Get Knowledge Base Stats",
        "description": "Get statistics for tenant's knowledge base.\n\nReturns document count and collection information.\n\nRequires admin role in JWT.",
        "operationId": "get_knowledge_base_stats_api_admin_tenants__tenant_id__knowledge_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DocumentIngestRequest": {
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Documents",
            "description": "List of document texts to ingest"
          },
          "metadatas": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadatas",
            "description": "Optional metadata for each document"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "DocumentIngestRequest",
        "description": "Request to ingest documents into knowledge base."
      },
      "DocumentIngestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Document Ids"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "collection_name",
          "document_count",
          "document_ids"
        ],
        "title": "DocumentIngestResponse",
        "description": "Response for document ingestion."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeBaseStatsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "collection_name",
          "document_count"
        ],
        "title": "KnowledgeBaseStatsResponse",
        "description": "Response for knowledge base statistics."
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "description": "Generic message response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}